Restructure into Julia package
