+++ README.md
@@ -1,3 +1,3 @@
-# LiSCPOMDP
-All of my work for my Summer 2023 REU and CS195 in SISL. Implementation of and solving a POMDP designed to help the U.S. choose optimal actions (mining, exploring) in their quest
-for self sufficiency in developing a lithium supply chain.
+# LiPOMDPs.jl
+
+The packaged version of the original [LiSCPOMDP by yalonso03](https://github.com/yalonso03/LiSCPOMDP).
