--- logs/eval_log.txt
@@ -1,40 +1,0 @@
-==========start 20 simulations for RandomPolicy==========
-
-RandomPolicy Results:
-rdisc mean: 21.814249567366907, stdev: 7.060665814049403
-edisc mean: -41.357347639552, stdev: 7.00521784571318
-rtot mean: 29.271000000000004, stdev: 9.412142717060542
-etot mean: -48.1, stdev: 6.6165661311907895
-vt mean: -0.35, stdev: 2.007223796297028
-==========start 20 simulations for EfficiencyPolicy==========
-
-EfficiencyPolicy Results:
-rdisc mean: 23.79300952560279, stdev: 0.5775129270205505
-edisc mean: -56.34297492789435, stdev: 3.0361509465558996
-rtot mean: 30.244500000000006, stdev: 0.5165317028024512
-etot mean: -62.4, stdev: 2.6832815729997486
-vt mean: -1.0, stdev: 0.0
-==========start 20 simulations for EfficiencyPolicyWithUncertainty==========
-
-EfficiencyPolicyWithUncertainty Results:
-rdisc mean: 23.270036307381122, stdev: 0.45441780424628897
-edisc mean: -57.92774225583882, stdev: 3.409166470114269
-rtot mean: 29.617499999999996, stdev: 0.3689512162874642
-etot mean: -64.3, stdev: 3.130495168499706
-vt mean: -1.0, stdev: 0.0
-==========start 20 simulations for EmissionAwarePolicy==========
-
-EmissionAwarePolicy Results:
-rdisc mean: 25.826343745294547, stdev: 0.7268820787222737
-edisc mean: -50.181356080343576, stdev: 2.5835171535203756
-rtot mean: 32.52149999999999, stdev: 0.6641121893174373
-etot mean: -55.5, stdev: 2.23606797749979
-vt mean: -1.0, stdev: 0.0
-==========start 20 simulations for POMCPOWPlanner{LiPOMDP, POMCPOW.POWNodeFilter, MaxUCB, RandomActionGenerator{Random._GLOBAL_RNG}, typeof(estimate_value), Int64, Float64, POMCPOWSolver{Random._GLOBAL_RNG, POMCPOW.var"#6#12"}}==========
-
-POMCPOWPlanner{LiPOMDP, POMCPOW.POWNodeFilter, MaxUCB, RandomActionGenerator{Random._GLOBAL_RNG}, typeof(estimate_value), Int64, Float64, POMCPOWSolver{Random._GLOBAL_RNG, POMCPOW.var"#6#12"}} Results:
-rdisc mean: 23.729578168967336, stdev: 1.367048728720891
-edisc mean: -38.50182748690302, stdev: 3.4085955782564574
-rtot mean: 30.954, stdev: 1.2604735952055977
-etot mean: -44.45, stdev: 2.855742137850464
-vt mean: -1.0, stdev: 0.0
