--- logs/log_eval.txt
@@ -1,40 +1,0 @@
-==========start 100 simulations for RandomPolicy==========
-
-RandomPolicy Results:
-rdisc mean: 0.018025762426339582, stdev: 5.282454287417558
-edisc mean: -24.784363844596413, stdev: 9.979823052949715
-rtot mean: 0.4587000000000003, stdev: 6.922280621298157
-etot mean: -29.49, stdev: 11.584991161067851
-vt mean: 4.99, stdev: 1.8394937819240853
-==========start 100 simulations for EfficiencyPolicy==========
-
-EfficiencyPolicy Results:
-rdisc mean: 24.045068337538304, stdev: 0.259790831534361
-edisc mean: -56.43534006307913, stdev: 1.3532043479753344
-rtot mean: 30.666900000000005, stdev: 0.23099999999999984
-etot mean: -61.88, stdev: 1.2
-vt mean: 10.98, stdev: 0.19999999999999998
-==========start 100 simulations for EfficiencyPolicyWithUncertainty==========
-
-EfficiencyPolicyWithUncertainty Results:
-rdisc mean: -10.267727582655471, stdev: 3.132240300462001
-edisc mean: -65.15670562039617, stdev: 2.010187182528386
-rtot mean: -11.487300000000005, stdev: 3.926999999999999
-etot mean: -74.78, stdev: 2.2
-vt mean: 3.06, stdev: 0.5999999999999998
-==========start 100 simulations for EmissionAwarePolicy==========
-
-EmissionAwarePolicy Results:
-rdisc mean: 28.95932410666377, stdev: 0.29588699209298225
-edisc mean: -41.543655914214085, stdev: 1.2438220432522384
-rtot mean: 35.616899999999994, stdev: 0.2310000000000003
-etot mean: -46.88, stdev: 1.2
-vt mean: 10.98, stdev: 0.19999999999999998
-==========start 100 simulations for POMCPOWPlanner{LiPOMDP, POMCPOW.POWNodeFilter, MaxUCB, RandomActionGenerator{Random._GLOBAL_RNG}, typeof(estimate_value), Int64, Float64, POMCPOWSolver{Random._GLOBAL_RNG, POMCPOW.var"#6#12"}}==========
-
-POMCPOWPlanner{LiPOMDP, POMCPOW.POWNodeFilter, MaxUCB, RandomActionGenerator{Random._GLOBAL_RNG}, typeof(estimate_value), Int64, Float64, POMCPOWSolver{Random._GLOBAL_RNG, POMCPOW.var"#6#12"}} Results:
-rdisc mean: 6.955475212939996, stdev: 14.219311881473356
-edisc mean: -36.40581371618271, stdev: 14.466635547297322
-rtot mean: 9.1806, stdev: 17.6283134757696
-etot mean: -42.54, stdev: 17.00767503514252
-vt mean: 6.13, stdev: 1.9676936199058486
