--- logs/main_log.txt
@@ -1,42 +1,0 @@
-POMCPOW Solver
-==========start 10 simulations for RandomPolicy==========
-
-RandomPolicy Results:
-rdisc mean: -0.45340525758992756, stdev: 0.92967260459596
-edisc mean: -11.174190173988572, stdev: 7.434930612376062
-rtot mean: -0.42900000000000016, stdev: 1.2738170983308397
-etot mean: -13.5, stdev: 8.834905520465714
-vt mean: 2.6, stdev: 1.577621275493231
-==========start 10 simulations for EfficiencyPolicy==========
-
-EfficiencyPolicy Results:
-rdisc mean: 27.653594670533284, stdev: 0.8968456150654985
-edisc mean: -43.57400741845922, stdev: 4.050980993827471
-rtot mean: 34.41899999999998, stdev: 0.730486139498892
-etot mean: -48.8, stdev: 3.7947331922020555
-vt mean: 10.8, stdev: 0.6324555320336758
-==========start 10 simulations for EfficiencyPolicyWithUncertainty==========
-
-EfficiencyPolicyWithUncertainty Results:
-rdisc mean: -10.231753062934882, stdev: 11.676635491650703
-edisc mean: -62.662966625755345, stdev: 8.521572115050068
-rtot mean: -11.253000000000002, stdev: 14.505367627192356
-etot mean: -72.0, stdev: 9.486832980505138
-vt mean: 2.7, stdev: 2.213594362117866
-==========start 10 simulations for EmissionAwarePolicy==========
-
-EmissionAwarePolicy Results:
-rdisc mean: 28.709797855220092, stdev: 0.9415699196924029
-edisc mean: -40.37339170728706, stdev: 3.9154527979883635
-rtot mean: 35.40899999999999, stdev: 0.7304861394988965
-etot mean: -45.8, stdev: 3.7947331922020555
-vt mean: 9.0, stdev: 0.0
-==========start 10 simulations for POMCPOWPlanner{LiPOMDP, POMCPOW.POWNodeFilter, MaxUCB, RandomActionGenerator{Random._GLOBAL_RNG}, typeof(estimate_value), Int64, Float64, POMCPOWSolver{Random._GLOBAL_RNG, POMCPOW.var"#6#12"}}==========
-
-POMCPOWPlanner{LiPOMDP, POMCPOW.POWNodeFilter, MaxUCB, RandomActionGenerator{Random._GLOBAL_RNG}, typeof(estimate_value), Int64, Float64, POMCPOWSolver{Random._GLOBAL_RNG, POMCPOW.var"#6#12"}} Results:
-rdisc mean: 8.271774896490491, stdev: 11.981521225176854
-edisc mean: -42.71891156978496, stdev: 9.690093360953277
-rtot mean: 11.286000000000001, stdev: 15.432023846534195
-etot mean: -50.1, stdev: 10.461038826681284
-vt mean: 7.2, stdev: 1.7511900715418263
-pomcpow policy saved as planners/pomcpow_planner.jld2
