+++ src/LiPOMDPs.jl
@@ -1,0 +1,44 @@
+module LiPOMDPs
+
+using Random
+using POMDPs
+using POMCPOW
+using POMDPModels
+using POMDPTools
+using POMDPModelTools
+using POMDPPolicies
+using Serialization
+using Distributions
+using ParticleFilters
+using Parameters
+using Plots
+using LinearAlgebra
+using Statistics
+using D3Trees
+ 
+export 
+    #Abstract types
+    LiPOMDP, #<: POMDP
+    LiBelief #<: UnivariateDistribution
+include("LiPOMDP.jl")
+
+export
+    #Functions
+    evaluate_policies,
+    evaluate_policy,
+    print_policy_results,
+    replicate_simulation,
+    simulate_policy
+include("utils.jl")
+
+
+export 
+    EfficiencyPolicy
+
+include("policies.jl")
+
+
+
+
+
+end #module
